user cannot operate on his playlists when inactive

+++ src/main/java/com/musicplaylists/demo/musicplaylists/controllers/PlaylistController.java
@@ -2,10 +2,7 @@ package com.musicplaylists.demo.musicplaylists.controllers;
 
 import com.musicplaylists.demo.musicplaylists.dtos.PlaylistCreationDTO;
 import com.musicplaylists.demo.musicplaylists.dtos.PlaylistCreationResponseDTO;
-import com.musicplaylists.demo.musicplaylists.services.PlaylistService;
-import com.musicplaylists.demo.musicplaylists.services.ProfileUtils;
-import com.musicplaylists.demo.musicplaylists.services.SongService;
-import com.musicplaylists.demo.musicplaylists.services.ValidationUtils;
+import com.musicplaylists.demo.musicplaylists.services.*;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
@@ -16,7 +13,10 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.HttpClientErrorException;
+import org.springframework.web.server.ResponseStatusException;
 
+import javax.swing.*;
 import java.util.List;
 
 @RestController
@@ -26,12 +26,14 @@ public class PlaylistController {
     private final PlaylistService playlistService;
     private final SongService songService;
     private final ProfileUtils profileUtils;
+    private final UserService userService;
 
     @Autowired
-    public PlaylistController(PlaylistService playlistService, SongService songService, ProfileUtils profileUtils) {
+    public PlaylistController(PlaylistService playlistService, SongService songService, ProfileUtils profileUtils, UserService userService) {
         this.playlistService = playlistService;
         this.songService = songService;
         this.profileUtils = profileUtils;
+        this.userService = userService;
     }
 
     @Operation(summary = "Create a new playlist")
@@ -40,7 +42,11 @@ public class PlaylistController {
             @ApiResponse(responseCode = "400", description = "Validation errors occurred")
     })
     @PostMapping
-    public ResponseEntity<?> createPlaylist(@RequestBody @Valid PlaylistCreationDTO playlistDTO, BindingResult bindingResult) {
+    public ResponseEntity<?> createPlaylist(@RequestBody @Valid PlaylistCreationDTO playlistDTO, BindingResult bindingResult, Authentication authentication) {
+
+        if (!userService.checkIfUserActive(authentication)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized");
+        }
 
         return ValidationUtils.handleValidationResult(bindingResult, () -> {
             PlaylistCreationResponseDTO responseDTO = playlistService.createPlaylist(playlistDTO);
@@ -56,6 +62,10 @@ public class PlaylistController {
     @PostMapping("/song/{songId}")
     public ResponseEntity<?> addSongToPlaylists(@PathVariable Long songId, @RequestBody List<Long> playlistIds, Authentication authentication) {
 
+        if (!userService.checkIfUserActive(authentication)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized");
+        }
+
         songService.addSongToPlaylists(songId, playlistIds, authentication.getName());
         return new ResponseEntity<>(HttpStatus.OK);
     }
@@ -68,6 +78,10 @@ public class PlaylistController {
     @DeleteMapping("/{playlistId}")
     public ResponseEntity<?> deletePlaylist(@PathVariable @Valid Long playlistId, Authentication authentication) {
 
+        if (!userService.checkIfUserActive(authentication)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized");
+        }
+
         playlistService.deletePlaylist(playlistId, authentication.getName());
         return new ResponseEntity<>(HttpStatus.OK);
     }
@@ -79,6 +93,10 @@ public class PlaylistController {
     })
     @DeleteMapping("/song/{songId}")
     public ResponseEntity<?> removeSongFromPlaylists(@PathVariable Long songId, @RequestBody List<Long> playlistIds, Authentication authentication) {
+
+        if (!userService.checkIfUserActive(authentication)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized");
+        }
 
         songService.removeSongFromPlaylists(songId, playlistIds, authentication.getName());
         return new ResponseEntity<>(HttpStatus.OK);
